Add wak_math library for computing the largest prime factor of a given number.
Introduces a new library `wak_math` with the function `get_largest_prime_factor` to determine the largest prime factor of a number. The addition includes the necessary CMake setup for building the library and its integration into the project. This function utilizes an efficient algorithm to find the largest prime factor by iterating up to the square root of the input number.