Add summation of primes functionality
Introduce a new function `sum_primes` in the wak_math library to calculate the sum of prime numbers up to a specified limit. Create a CMake configuration to build the summation_of_primes executable, linking it with the wak_math library. This update enables computing the sum of all primes below 2,000,000.